+++ spring-app/src/main/kotlin/pl/starchasers/up/service/UserService.kt
@@ -6,11 +6,11 @@ import pl.starchasers.up.exception.UserException
 import pl.starchasers.up.repository.UserRepository
 
 interface UserService {
-    fun getUser(id: Int): User
+    fun getUser(id: Long): User
 
     fun getUser(username: String): User
 
-    fun findUser(id: Int): User?
+    fun findUser(id: Long): User?
 
     fun findUser(username: String): User?
 }
@@ -19,13 +19,13 @@ interface UserService {
 class UserServiceImpl(
         private val userRepository: UserRepository
 ) : UserService {
-    override fun getUser(id: Int): User = userRepository.findFirstById(id)
+    override fun getUser(id: Long): User = userRepository.findFirstById(id)
             ?: throw UserException("User with ID `$id` doesn't exist")
 
     override fun getUser(username: String): User = userRepository.findFirstByUsername(username)
             ?: throw UserException("User with username '$username' doesn't exist")
 
-    override fun findUser(id: Int): User? = userRepository.findFirstById(id)
+    override fun findUser(id: Long): User? = userRepository.findFirstById(id)
 
     override fun findUser(username: String): User? = userRepository.findFirstByUsername(username)
